+++ Rdmp.Dicom/PipelineComponents/DicomSources/Worklists/IDicomFileWorklist.cs
@@ -1,4 +1,5 @@
 using System.IO;
+using Rdmp.Dicom.Extraction.FoDicomBased;
 
 namespace Rdmp.Dicom.PipelineComponents.DicomSources.Worklists
 {
@@ -8,6 +9,6 @@ namespace Rdmp.Dicom.PipelineComponents.DicomSources.Worklists
         /// Populates a DirectoryInfo or FileInfo depending on what the next dicom file system collection to process is (or returns false if no more processing is required)
         /// </summary>
         /// <returns></returns>
-        bool GetNextFileOrDirectoryToProcess(out DirectoryInfo directory, out FileInfo file);
+        bool GetNextFileOrDirectoryToProcess(out DirectoryInfo directory, out AmbiguousFilePath file);
     }
 }